test(rsbuild-plugin-angular): add utils
